+++ lib/plugins/index.js
@@ -11,7 +11,7 @@ module.exports = [
   require('./info.js'),
   require('./logs.js'),
   require('./metrics.js'),
-  require('./print/print.js'),
+  require('./print.js'),
   require('./remove/remove.js'),
   require('./rollback/index.js'),
   require('./slstats/slstats.js'),
