+++ lib/plugins/print.js
@@ -3,8 +3,7 @@
 const os = require('os');
 const _ = require('lodash');
 const BbPromise = require('bluebird');
-const getServerlessConfigFile = require('../../utils/getServerlessConfigFile')
-  .getServerlessConfigFile;
+const getServerlessConfigFile = require('../utils/getServerlessConfigFile').getServerlessConfigFile;
 const jc = require('json-cycle');
 const YAML = require('js-yaml');
